Combined JPEG2K codecs to allow combination use
