+++ core/include/codec/codec_jpeg2K.h
@@ -73,15 +73,13 @@ JPEG2K_EXTERN_DECL void (*opj_destroy_codec)(opj_codec_t *);
 // Maximum number of components within an image allowed
 #define NUM_COMPS_MAX 4
 
-class CodecJPEG2K : public Codec {
+class CodecJPEG2K_base : public Codec {
  public:
   
-  CodecJPEG2K(int compression_level, int64_t* tile_dims):Codec(compression_level) {
+  CodecJPEG2K_base(int compression_level, int64_t* tile_dims):Codec(compression_level) {
     static bool loaded = false;
     static std::mutex loading;
     
-    tile_image_width_ = tile_dims[0];
-    tile_image_height_ = tile_dims[1];
 
     if (!loaded) {
       loading.lock();
@@ -134,14 +132,36 @@ class CodecJPEG2K : public Codec {
       }
     }
   }
-  
-  int compress_tile(unsigned char* tile, size_t tile_size, void** tile_compressed, size_t& tile_compressed_size);
+};
 
-  int decompress_tile(unsigned char* tile_compressed,  size_t tile_compressed_size, unsigned char* tile, size_t tile_size);
+class CodecJPEG2K : virtual public CodecJPEG2K_base {
+
+public: 
+  CodecJPEG2K(int compression_level, int64_t* tile_dims):CodecJPEG2K_base(compression_level, tile_dims) { 
+     tile_image_width_ = tile_dims[0];
+     tile_image_height_ = tile_dims[1];
+  }
+  
+   int compress_tile(unsigned char* tile, size_t tile_size, void** tile_compressed, size_t& tile_compressed_size);
+
+   int decompress_tile(unsigned char* tile_compressed,  size_t tile_compressed_size, unsigned char* tile, size_t tile_size);
   
 private:
-  int64_t tile_image_width_;
-  int64_t tile_image_height_;
+   int64_t tile_image_width_;
+   int64_t tile_image_height_;
+
+};
+
+
+class CodecJPEG2K_RGB : virtual public CodecJPEG2K_base {
+
+public:
+   CodecJPEG2K_RGB(int compression_level, int64_t* tile_dims):CodecJPEG2K_base(compression_level, tile_dims) { }
+  
+   int compress_tile(unsigned char* tile, size_t tile_size, void** tile_compressed, size_t& tile_compressed_size);
+
+   int decompress_tile(unsigned char* tile_compressed,  size_t tile_compressed_size, unsigned char* tile, size_t tile_size);
+
 };
 
 #endif /*__CODEC_JPEG200_H__*/
