+++ core/src/codec/codec_jpeg2K.cc
@@ -27,17 +27,17 @@
  * 
  * @section DESCRIPTION
  *
- * This file implements codec for OpenJPEG 2000 compression and decompression.
+ * This file implements codecs for OpenJPEG 2000 compression and decompression.
  */
 
-#ifdef ENABLE_JPEG2K
-
+#if defined(ENABLE_JPEG2K) || defined(ENABLE_JPEG2K_RGB)
 #define JPEG2K_EXTERN_DECL extern
 #include "codec_jpeg2K.h"
 
+/** Debugging functions -- once algorithms are working, should not be needed **/
+
 void print_POC(opj_poc_t POC)
 {
-
 /**
  * Progression order changes
  *
@@ -209,7 +209,7 @@ void print_image_info (opj_image_t *I)
    }
 }
 
-
+/** END Debugging functions **/
 
 void cleanup(void *l_data, opj_stream_t *l_stream, opj_codec_t *l_codec, opj_image_t *l_image)
 {
@@ -237,6 +237,22 @@ void get_header_values(unsigned char *tile_in,
 
    return;
 }
+
+#endif
+
+#ifdef ENABLE_JPEG2K
+/**
+ * 
+ * @section DESCRIPTION
+ *
+ * This class implements codec for OpenJPEG 2000 compression and decompression.
+ *   This codec will compress a single grap-scale component image. For 
+ *   multiple component images, each individual component section should be
+ *   assigned to a separate attribute and tiled as individual at the level of 
+ *   individual pixels.
+ *
+ *   FUTURE enhancement: May need to find method to pass bits per pixel.
+ */
 
 int CodecJPEG2K::compress_tile(unsigned char* tile_in, size_t tile_size_in, void** tile_compressed, size_t& tile_compressed_size) 
 {
@@ -507,6 +523,360 @@ int CodecJPEG2K::decompress_tile(unsigned char* tile_compressed, size_t tile_com
 
    // Add header portion -- write ints into char array
    OPJ_UINT32 *buffer = (OPJ_UINT32 *)tile_out;
+   OPJ_BYTE *ob = l_data;
+
+   // Copy pixel data to output buffer
+   size_t i;
+   for (i = 0; i < l_data_size; ++i) {
+      buffer[i] = (OPJ_UINT32) (0x000000FF & (*ob));
+      ++ob;
+   }
+
+   //memcpy(ob, l_data, copy_bytes);
+
+   /* Free memory */
+   cleanup(l_data, l_stream, l_codec, l_image);
+
+   // Success
+   return TILEDB_CD_OK;
+}
+
+#endif
+
+
+/**
+ * 
+ * @section DESCRIPTION
+ *
+ * This class implements codec for OpenJPEG 2000 compression and decompression.
+ *   Each individual tile compressed will be treated as a separate image with
+ *   three componenets (RGB) and the raw pixel data will be prefaced with a
+ *   "header" of 3 integers: number of components (3), image width, 
+ *   image height. These tiles will need to be "hand" created; the TileDB
+ *   tile will then correspond to a blob of RGB pixels and prefaced header.
+ *
+ *   FUTURE enhancement: pack bits per pixel into upper 16 bits of num_comps
+ *     To be used for images with more than 8 bpp (or for int16_t pixels).
+ */
+
+#ifdef ENABLE_JPEG2K_RGB
+
+// Number of int32_t that will be in header
+#define HEADER_SIZE 3
+
+int CodecJPEG2K_RGB::compress_tile(unsigned char* tile_in, size_t tile_size_in, void** tile_compressed, size_t& tile_compressed_size) 
+{
+   opj_cparameters_t l_param;
+   opj_codec_t * l_codec;
+   opj_image_t * l_image;
+   opj_image_cmptparm_t l_image_params[NUM_COMPS_MAX];
+   opj_stream_t * l_stream;
+   //OPJ_UINT32 l_nb_tiles_width, l_nb_tiles_height, l_nb_tiles;
+   //OPJ_UINT32 l_data_size;
+   //size_t len;
+
+   opj_image_cmptparm_t * l_current_param_ptr;
+   OPJ_UINT32 i;
+   OPJ_BYTE *l_data;
+
+   OPJ_UINT32 num_comps, compno;
+   int image_width;
+   int image_height;
+   //int tile_width;
+   //int tile_height;
+
+// Fixed parameter values. May need to amend if different image types are used
+   int comp_prec = 8;
+   int irreversible = 0; // 0 = lossless
+   int cblockw_init = 64;
+   int cblockh_init = 64;
+   int numresolution = 6;
+   OPJ_UINT32 offsetx = 0;
+   OPJ_UINT32 offsety = 0;
+   int quality_loss = 0; // lossless
+
+   get_header_values(tile_in, &num_comps, &image_width, &image_height);
+
+   opj_set_default_encoder_parameters(&l_param); // Default parameters
+
+   if (num_comps > NUM_COMPS_MAX) {
+      char msg[80];
+      sprintf(msg, "ERROR -> j2k_compress: num_comps %d > NUM_COMPS_MAX %d\n", num_comps, NUM_COMPS_MAX);
+      return print_errmsg(msg);
+   }
+
+//  Set data pointer to the first pixel byte beyond the header values
+   OPJ_UINT32 header_offset = 3 * sizeof(int);
+   l_data = (OPJ_BYTE*) &tile_in[header_offset];
+
+   /** number of quality layers in the stream */
+   if (quality_loss) {
+      l_param.tcp_numlayers = 1;
+      l_param.cp_fixed_quality = 1;
+      l_param.tcp_distoratio[0] = 20;
+   }
+   else { // Set for lossless compression
+      l_param.cp_disto_alloc = 1;
+      l_param.cp_fixed_alloc = 0;
+      l_param.tcp_numlayers = 1;
+      l_param.tcp_rates[0] = 0.0;
+      l_param.tcp_distoratio[0] = 0.0;
+   }
+
+   /* tile definitions parameters */
+   /* position of the tile grid aligned with the image */
+   l_param.cp_tx0 = 0;
+   l_param.cp_ty0 = 0;
+
+   l_param.tile_size_on = OPJ_FALSE;
+   l_param.cp_tdx = 0;
+   l_param.cp_tdy = 0;
+
+   /* code block size */
+   l_param.cblockw_init = cblockw_init;
+   l_param.cblockh_init = cblockh_init;
+
+   /* use irreversible encoding ?*/
+   l_param.irreversible = irreversible;
+
+   /** number of resolutions */
+   l_param.numresolution = numresolution;
+
+   /** progression order to use*/
+   l_param.prog_order = OPJ_LRCP; // default
+
+    /** Multiple component transform? **/
+    if (num_comps == 1)
+       l_param.tcp_mct = 0;
+    else
+       l_param.tcp_mct = 1; /* multiple components in image */
+
+   /* image definition */
+   l_current_param_ptr = l_image_params;
+   for (i = 0; i < num_comps; ++i) {
+       /* do not bother bpp useless */
+       l_current_param_ptr->bpp = 8;     // Does this need to be parameter?
+       l_current_param_ptr->dx = 1;
+       l_current_param_ptr->dy = 1;
+
+       l_current_param_ptr->w = (OPJ_UINT32)image_width;
+       l_current_param_ptr->h = (OPJ_UINT32)image_height;
+
+       l_current_param_ptr->sgnd = 0;
+       l_current_param_ptr->prec = (OPJ_UINT32)comp_prec;
+
+       l_current_param_ptr->x0 = offsetx;
+       l_current_param_ptr->y0 = offsety;
+
+       ++l_current_param_ptr;
+   }
+
+   /* should we do j2k or jp2 ?*/
+   l_codec = opj_create_compress(OPJ_CODEC_JP2);
+   // l_codec = opj_create_compress(OPJ_CODEC_J2K);
+
+   if (! l_codec) {
+      char msg[100];
+      sprintf(msg, "ERROR -> j2k_compress: failed to setup codec!\n");
+      return print_errmsg(msg);
+   }
+
+   if (num_comps == 3)
+      l_image = opj_image_create(num_comps, l_image_params, OPJ_CLRSPC_SRGB);
+   else if (num_comps == 1)
+      l_image = opj_image_create(num_comps, l_image_params, OPJ_CLRSPC_GRAY);
+   else
+      l_image = opj_image_create(num_comps, l_image_params, OPJ_CLRSPC_UNKNOWN);
+
+   if (! l_image) {
+      cleanup(NULL, NULL, l_codec, NULL);
+      char msg[100];
+      sprintf(msg, "ERROR -> j2k_compress: failed to setup tile image!\n");
+      return print_errmsg(msg);
+   }
+
+   l_image->x0 = offsetx;
+   l_image->y0 = offsety;
+   l_image->x1 = offsetx + (OPJ_UINT32)image_width;
+   l_image->y1 = offsety + (OPJ_UINT32)image_height;
+/** should be set in opj_image_create() **
+   if (num_comps == 3)
+      l_image->color_space = OPJ_CLRSPC_SRGB;
+   else if (num_comps == 1)
+      l_image->color_space = OPJ_CLRSPC_GRAY;
+   else
+      l_image->color_space = OPJ_CLRSPC_UNKNOWN;
+**/
+
+// Copy pixels to l_image component data from l_data (tile_in)
+
+   size_t num_pixels = image_width * image_height;
+   size_t num_bytes = num_pixels * sizeof(OPJ_INT32);
+   for (compno = 0; compno < num_comps; ++compno) {
+      memcpy(l_image->comps[compno].data, l_data, num_bytes);
+      l_data += num_bytes;
+   }
+
+// CPB - to debug by comparing image information and parameters from good run
+//print_image_info(l_image);
+//print_parameters(l_param);
+// CPB end
+
+   if (! opj_setup_encoder(l_codec, &l_param, l_image)) {
+      cleanup(NULL, NULL, l_codec, l_image);
+      return print_errmsg("jpeg2k_compress: failed to setup the codec");
+   }
+
+   l_stream = (opj_stream_t *) opj_stream_create_default_memory_stream(OPJ_FALSE);
+   if (! l_stream) {
+      cleanup(NULL, NULL, l_codec, l_image);
+      return print_errmsg("jpeg2k_compress: failed to create the memory stream");
+   }
+
+   if (! opj_start_compress(l_codec, l_image, l_stream)) {
+      cleanup(NULL, l_stream, l_codec, l_image);
+      return print_errmsg("jpeg2k_compress: failed to start compress");
+   }
+
+   if (! opj_encode(l_codec, l_stream)) {
+      cleanup(NULL, l_stream, l_codec, l_image);
+      return print_errmsg("jpeg2k_compress: failed to encode the tile");
+   }
+
+   if (! opj_end_compress(l_codec, l_stream)) {
+      cleanup(NULL, l_stream, l_codec, l_image);
+      return print_errmsg("jpeg2k_compress: failed to end compress");
+   }
+
+   *tile_compressed = opj_mem_stream_copy(l_stream, &tile_compressed_size);
+
+   cleanup(NULL, l_stream, l_codec, l_image);
+
+   // Success
+   return TILEDB_CD_OK;
+}
+
+int CodecJPEG2K_RGB::decompress_tile(unsigned char* tile_compressed, size_t tile_compressed_size, unsigned char* tile_out, size_t size_out) 
+{
+   opj_dparameters_t l_param;
+   opj_codec_t * l_codec;
+   opj_image_t * l_image;
+   opj_stream_t * l_stream;
+   OPJ_UINT32 l_data_size;
+   OPJ_UINT32 l_max_data_size = 1000;
+   OPJ_BYTE * l_data = (OPJ_BYTE *) malloc(1000);
+
+   OPJ_UINT32 l_tile_index;
+   OPJ_INT32 l_current_tile_x0;
+   OPJ_INT32 l_current_tile_y0;
+   OPJ_INT32 l_current_tile_x1;
+   OPJ_INT32 l_current_tile_y1;
+   OPJ_UINT32 l_nb_comps=0;
+   OPJ_BOOL l_go_on = OPJ_TRUE;
+
+   OPJ_UINT32 image_height;
+   OPJ_UINT32 image_width;
+   OPJ_UINT32 numcomps;
+
+   /* Set the default decoding parameters */
+   opj_set_default_decoder_parameters(&l_param);
+   l_param.decod_format = JP2_CFMT;
+
+   /** you may here add custom decoding parameters */
+   /* do not use layer decoding limitations */
+   l_param.cp_layer = 0;
+
+   /* do not use resolutions reductions */
+   l_param.cp_reduce = 0;
+
+   /** Fixed to match Compress format **/
+   l_codec = opj_create_decompress(OPJ_CODEC_JP2);
+   if (!l_codec){
+      cleanup(l_data, NULL, NULL, NULL);
+      return print_errmsg("jpeg2k_decompress: failed to create decompress codec");
+   }
+
+   /** Setup stream with user data into stream **/
+   l_stream = opj_stream_create_memory_stream((void *)tile_compressed, tile_compressed_size, OPJ_TRUE);
+   if (!l_stream){
+      cleanup(l_data, NULL, l_codec, NULL);
+      return print_errmsg("jpeg2k_decompress: failed to create the memory stream");
+   }
+
+   /* Setup the decoder decoding parameters using user parameters */
+   if (! opj_setup_decoder(l_codec, &l_param)) {
+      cleanup(l_data, l_stream, l_codec, NULL);
+      return print_errmsg("jpeg2k_decompress: failed to setup the decoder");
+   }
+
+   /* Read the main header of the codestream and if necessary the JP2 boxes*/
+   if (! opj_read_header(l_stream, l_codec, &l_image)) {
+      cleanup(l_data, l_stream, l_codec, NULL);
+      return print_errmsg("jpeg2k_decompress: failed to read the header");
+   }
+
+   // PULL OUT numcomps, image_height, image_width from  l_stream
+   image_height = l_image->comps[0].h;
+   image_width = l_image->comps[0].w;
+   numcomps = l_image->numcomps;
+
+   if (!opj_set_decode_area(l_codec, l_image, 0, 0, 0, 0)){ // whole image
+      cleanup(l_data, l_stream, l_codec, l_image);
+      return print_errmsg("jpeg2k_decompress: failed to set the decoded area");
+   }
+
+   if (! opj_read_tile_header( l_codec, l_stream, &l_tile_index, &l_data_size,
+                               &l_current_tile_x0, &l_current_tile_y0,
+                               &l_current_tile_x1, &l_current_tile_y1,
+                               &l_nb_comps, &l_go_on)) {
+      cleanup(l_data, l_stream, l_codec, l_image);
+      return print_errmsg("jpeg2k_decompress: failed to read_tile_header");
+   }
+
+   if (l_go_on) {
+      if (l_data_size > l_max_data_size) {
+         OPJ_BYTE *l_new_data = (OPJ_BYTE *) realloc(l_data, l_data_size+HEADER_SIZE*sizeof(int));
+         if (! l_new_data) {
+            cleanup(l_data, l_stream, l_codec, l_image);
+            char msg[64];
+            sprintf(msg, "jpeg2k_decompress: failed to realloc new data of size %d\n", l_data_size);
+            return print_errmsg(msg);
+         }
+         l_data = l_new_data;
+         l_max_data_size = l_data_size;
+      }
+
+      if (! opj_decode_tile_data(l_codec,l_tile_index,l_data,l_data_size,l_stream)) {
+         cleanup(l_data, l_stream, l_codec, l_image);
+         return print_errmsg("jpeg2k_decompress: unable to decode_tile_data");
+      }
+   }  // if (l_go_on) 
+   else {               // Problem in the opj_read_tile_header() 
+       cleanup(l_data, l_stream, l_codec, l_image);
+       return print_errmsg("jpeg2k_decompress: Current tile number and total number of tiles problem");
+   }
+
+   if (! opj_end_decompress(l_codec,l_stream)) {
+      cleanup(l_data, l_stream, l_codec, l_image);
+      return print_errmsg("jpeg2k_decompress: Failed to end_decompress correctly");
+   }
+
+   // Double check space in tile for decompressed data 
+   size_t image_bytes = l_data_size;
+   if (image_bytes > size_out) {
+      char msg[100];
+      sprintf(msg, "Tile decompresss size (%lu) greater than expected image size (%lu)", image_bytes, size_out);
+      return print_errmsg(msg);
+   }
+
+
+   // Add header portion -- write ints into char array
+   OPJ_UINT32 *buffer = (OPJ_UINT32 *)tile_out;
+   buffer[0] = numcomps;      // number of color components
+   buffer[1] = image_width;   // image width
+   buffer[2] = image_height;  // image height
+   buffer += 3;
+
    OPJ_BYTE *ob = l_data;
 
    // Copy pixel data to output buffer
